修改文件大小相关字段类型为长整型

+++ src/main/java/com/devteam/mikufunbackend/entity/Aria2FileV0.java
@@ -10,7 +10,7 @@ import lombok.Data;
 public class Aria2FileV0 {
     int index;
     String path;
-    int completedLength;
-    int length;
+    long completedLength;
+    long length;
     boolean selected;
 }
